删除多余业务代码

+++ KSWebRTC/Script/Chat/Interface/Chat/KSChatController.h
@@ -17,6 +17,6 @@
 /// @param isCaller 是否创建呼叫（主叫）
 /// @param peerId 对方id
 /// @param target 跳转控制器
-+ (void)callWithType:(KSCallType)type callState:(KSCallStateMaintenance)callState isCaller:(BOOL)isCaller peerId:(int)peerId target:(UIViewController *)target;
++ (void)callWithType:(KSCallType)type callState:(KSCallStateMaintenance)callState isCaller:(BOOL)isCaller peerId:(long long)peerId target:(UIViewController *)target;
 
 @end
